♻️ refactor :  RouteConfirmPanel 구조 정리
